# 01-Documents/API-SPEC-001-openapi.yaml
openapi: 3.0.3
info:
  title: RIS Invoice Discounting Platform API
  version: 1.0.0
  description: |
    REST API for the RIS Invoice Discounting and Early Payment Platform, Uganda.
    All endpoints require Bearer JWT authentication unless marked public.
    Role-based access is enforced at route and service layer independently.
  contact:
    name: RIS Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.mms.co.ug/v1
    description: Production
  - url: https://staging-api.mms.co.ug/v1
    description: Staging
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
          example: CREDIT_LIMIT_EXCEEDED
        message:
          type: string
          example: Invoice face value exceeds buyer credit limit
        details:
          type: object
          description: Field-level error details for validation failures
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            details:
              type: object
              additionalProperties:
                type: string
              example:
                face_value: Must be a positive integer
                due_date: Must be between 7 and 90 days from today
    MoneyAmount:
      type: integer
      format: int64
      description: Raw UGX integer — no decimal places. UGX has no subunits.
      example: 50000000
    UUID:
      type: string
      format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
...
    InvoiceStatus:
      type: string
      enum:
        [
          draft,
          submitted,
          buyer_confirmed,
          scored,
          approved,
          rejected,
          pending_first_auth,
          pending_second_auth,
          executing,
          funded,
          collecting,
          overdue,
          collected,
          defaulted,
          cancelled,
        ]
...
    InvoiceSubmission:
      type: object
      required: [invoice_number, buyer_id, face_value, due_date]
      properties:
        invoice_number:
          type: string
          maxLength: 100
          description: Must be unique per supplier
        buyer_id:
          $ref: '#/components/schemas/UUID'
        face_value:
          $ref: '#/components/schemas/MoneyAmount'
          description: Must be positive integer UGX
        due_date:
          type: string
          format: date
          description: Must be 7-90 days from today
        description:
          type: string
          maxLength: 500
    InvoiceResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        invoice_number:
          type: string
        buyer_id:
          $ref: '#/components/schemas/UUID'
        supplier_id:
          $ref: '#/components/schemas/UUID'
        face_value:
          $ref: '#/components/schemas/MoneyAmount'
        due_date:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/InvoiceStatus'
        sla_deadline:
          type: string
          format: date-time
        buyer_confirmed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Not authenticated or token expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error_code: UNAUTHORIZED
            message: Authentication required
    Forbidden:
      description: Authenticated but insufficient role or accessing another user's resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error_code: FORBIDDEN
            message: You do not have permission to access this resource
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationFailed:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error_code: RATE_LIMIT_EXCEEDED
            message: Too many requests. Please try again later.
    InternalError:
      description: Internal server error — no stack trace exposed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error_code: INTERNAL_ERROR
            message: An unexpected error occurred. Please contact support.

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /invoices/submit:
    post:
      summary: Submit invoice for discounting
      tags: [Invoices]
      x-required-role: supplier (owns invoice)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceSubmission'
      responses:
        '201':
          description: Invoice submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                supplier_not_approved:
                  value:
                    error_code: SUPPLIER_NOT_APPROVED
                    message: Your KYC is pending approval. You cannot submit invoices yet.
                duplicate_invoice:
                  value:
                    error_code: DUPLICATE_INVOICE
                    message: An invoice with this number already exists for your account.
                buyer_not_approved:
                  value:
                    error_code: BUYER_NOT_APPROVED
                    message: This buyer is not registered or is inactive.
                tenor_out_of_range:
                  value:
                    error_code: TENOR_OUT_OF_RANGE
                    message: Invoice tenor is 95 days. Must be between 7 and 90 days.
                    details:
                      calculated_tenor_days: 95
                credit_limit_exceeded:
                  value:
                    error_code: CREDIT_LIMIT_EXCEEDED
                    message: Invoice exceeds buyer credit limit.
                    details:
                      remaining_availability: 15000000
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalError'
  /invoices/{id}:
    get:
      summary: Get invoice by ID
      tags: [Invoices]
      x-required-role: supplier (own only), credit_officer, management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...